refactor: move away from services dir
